requirements.txt
gradio>=4.20.0,<5.0.0
sqlalchemy==2.0.25
transformers==4.36.2
torch==2.1.2
huggingface-hub==0.20.2
python-dotenv==1.0.0
pydantic==2.5.3
markdown==3.5.2
Pillow==10.2.0
